Rename to Dimension Hopper Tweaks and fix package name
You don't own thecomputerizer.com AND uppercase package names are meh
