--- src/main/java/com/TheComputerizer/DimensionalHopperFinalBoss/util/interfaces/IHasModel.java
@@ -1,5 +1,0 @@
-package com.TheComputerizer.DimensionalHopperFinalBoss.util.interfaces;
-
-public interface IHasModel {
-    public void registerModels();
-}
