+++ src/main/java/mods/thecomputerizer/dimensionhoppertweaks/entity/EntityFinalBoss.java
@@ -1,4 +1,4 @@
-package com.TheComputerizer.DimensionalHopperFinalBoss.entity;
+package mods.thecomputerizer.dimensionhoppertweaks.entity;
 
 import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.SharedMonsterAttributes;
