+++ src/main/java/mods/thecomputerizer/dimensionhoppertweaks/entity/model/ModelFinalBoss.java
@@ -1,4 +1,4 @@
-package com.TheComputerizer.DimensionalHopperFinalBoss.entity.model;
+package mods.thecomputerizer.dimensionhoppertweaks.entity.model;
 
 import net.minecraft.client.model.ModelBase;
 import net.minecraft.client.model.ModelRenderer;
@@ -50,7 +50,7 @@ public class ModelFinalBoss extends ModelBase {
     }
 
     @Override
-    public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5) { 
+    public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5) {
         this.body.render(f5);
         this.brain.render(f5);
         this.body_1.render(f5);
