+++ src/main/java/mods/thecomputerizer/dimensionhoppertweaks/util/interfaces/IHasModel.java
@@ -1,0 +1,5 @@
+package mods.thecomputerizer.dimensionhoppertweaks.util.interfaces;
+
+public interface IHasModel {
+    public void registerModels();
+}
