+++ src/main/resources/assets/dimensionhoppertweaks/textures/entity/finalboss.png
