+++ src/main/resources/mcmod.info
@@ -1,16 +1,16 @@
 [
-{
-  "modid": "dimensionalhopperfinalboss",
-  "name": "Dimension Hopper Final Boss",
-  "description": "Final Boss for my modpack Dimension Hopper",
-  "version": "1.0",
-  "mcversion": 1.12.2",
-  "url": "",
-  "updateUrl": "",
-  "authorList": ["The_Computerizer"],
-  "credits": "The_Computerizer",
-  "logoFile": "",
-  "screenshots": [],
-  "dependencies": []
-}
+  {
+    "modid": "dimensionhoppertweaks",
+    "name": "Dimension Hopper Tweaks",
+    "description": "Various tweaks and additions for my modpack Dimension Hopper",
+    "version": "1.0",
+    "mcversion": "1.12.2",
+    "url": "",
+    "updateUrl": "",
+    "authorList": ["The_Computerizer"],
+    "credits": "The_Computerizer",
+    "logoFile": "",
+    "screenshots": [],
+    "dependencies": []
+  }
 ]
